requirements.txt
Django>=5.2
Pillow
djangorestframework
